requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.14.0
python-dotenv>=1.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
chromadb>=0.4.22
